Revert "[refactor/#2]:코드 최적화"
This reverts commit 06dcd3a0fddaf29a3b5a0befc11dfcc3084345fc.
