+++ src/features/Auth/components/Join/ConsentForm.tsx
@@ -1,20 +1,75 @@
 import styled from 'styled-components';
-import {} from '@/constants/icons';
-import { TOS } from '../../constants/tos';
-import TermItem, { CheckIcon } from './ConsentTerm';
+import { ICONS } from '@/constants/icons';
+import { TOS, TOSProps } from '../../constants/tos';
+
+const CheckIcon = styled.img.attrs({ src: ICONS.check, alt: 'check' })`
+  width: 20px;
+  height: 20px;
+  fill: ${({ theme }) => theme.color.neutral[400]};
+`;
+
+function TermItem(term: TOSProps) {
+  const { id, title } = term;
+  const isOptional = id === 3;
+  return (
+    <Term>
+      <Option>
+        <Icon>
+          <CheckIcon />
+        </Icon>
+        <Essential>
+          <span> {isOptional ? '(선택)' : '(필수)'}</span>
+          {title}
+        </Essential>
+        {id ? <View>보기</View> : ''}
+      </Option>
+    </Term>
+  );
+}
+
+const Term = styled.div`
+  padding: 10px;
+`;
+const Icon = styled.div``;
+
+const Essential = styled.div`
+  margin-left: 3px;
+  color: ${({ theme }) => theme.color.neutral[400]};
+  ${({ theme }) => theme.typo.body.medium[14]}
+  span {
+    margin-right: 3px;
+    color: ${({ theme }) => theme.color.neutral[900]};
+    ${({ theme }) => theme.typo.body.semi_bold[14]}
+  }
+`;
+
+const Option = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 10px;
+`;
+
+const View = styled.span`
+  ${({ theme }) => theme.typo.body.semi_bold[14]};
+  color: ${({ theme }) => theme.color.neutral[600]};
+  border-bottom: 2px solid ${({ theme }) => theme.color.neutral[600]};
+  margin-left: 6px;
+`;
 
 function ConsentForm() {
   return (
     <Container>
       <AllAgree>
-        <CheckIcon />
+        <Icon>
+          <CheckIcon />
+        </Icon>
         모두 동의
       </AllAgree>
       <ul>
         {TOS.map((term) => {
           return (
-            <li key={term.id}>
-              <TermItem {...term} />
+            <li>
+              <TermItem key={term.id} {...term} />
             </li>
           );
         })}
@@ -32,14 +87,13 @@ const Container = styled.div`
     gap: 10px;
   }
 `;
-
 const AllAgree = styled.div`
-  margin-bottom: 20px;
-  padding: 18.5px 24px;
   display: flex;
   align-items: center;
   gap: 6px;
   color: ${({ theme }) => theme.color.neutral[400]};
-  ${({ theme }) => theme.typo.body.medium[14]};
+  ${({ theme }) => theme.typo.body.medium[14]}
+  padding: 18.5px 24px;
   border: 0.8px solid ${({ theme }) => theme.color.neutral[400]};
+  margin-bottom: 20px;
 `;
